requirements.txt
fastmcp==2.11.2
Flask==3.1.1
gradio==5.41.1
langchain_community==0.3.27
langchain_core==0.3.74
langchain_openai==0.3.28
langgraph==0.6.4
langgraph_supervisor==0.0.29
openai==1.99.4
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
requests==2.31.0
rich==14.1.0
structlog==25.4.0
typing_extensions==4.14.1
chromadb==0.5.5
pypdf==4.0.1
langchain-chroma==0.1.2
beautifulsoup4==4.12.2
